MODIFIED : handle for single pages and double pages
