+++ scripts/make_html.js
@@ -1,5 +1,6 @@
 const doPdf = require('../src/index');
-const makePdf = require('./pdf-lib_tests');
+const makePdf = require('./generate_pdf');
+const { PageSizes } = require('pdf-lib');
 const { program } = require('commander');
 const PythonShell = require('python-shell').PythonShell;
 
@@ -22,6 +23,21 @@ const keysOpt = Object.keys(options);
 if (!keysOpt.includes("config") && !keysOpt.includes("output")) {
     console.log(`You must give a config file (-c, --config) and an output file name (-o, --output)!\n${usage}`);
     process.exit(1);
+}
+
+const PAGE_SIZES = {
+    "A0" : PageSizes.A0,
+    "A1" : PageSizes.A1,
+    "A2" : PageSizes.A2,
+    "A3" : PageSizes.A3,
+    "A4" : PageSizes.A4,
+    "A5" : PageSizes.A5,
+    "A6" : PageSizes.A6,
+    "A7" : PageSizes.A7,
+    "LETTER" : PageSizes.Letter,
+    "letter" : PageSizes.Letter,
+    "EXECUTIVE" : [504.0, 720.0],
+    "EXECUTIVE_LULU_WITH_BLEED" : [521.57, 737.0]
 }
 
 const configPath = options.config;
@@ -50,7 +66,7 @@ doPdf({configPath, outputDirName, cliBookCode}).then(() => {
             console.log('finished');
         });
     } else {
-        makePdf(outputDirName);
+        makePdf({dirName:outputDirName, pageSize:PAGE_SIZES["EXECUTIVE_LULU_WITH_BLEED"]});
     }
 }).catch((e) => {
     console.error(e);
